Add support for `transient` property in workspaces and prevent window reloads (#119695)
