+++ src/vs/platform/workspaces/electron-main/workspacesManagementMainService.ts
@@ -3,15 +3,15 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { toWorkspaceFolders, IWorkspaceIdentifier, hasWorkspaceFileExtension, UNTITLED_WORKSPACE_NAME, IResolvedWorkspace, IStoredWorkspaceFolder, isStoredWorkspaceFolder, IWorkspaceFolderCreationData, IUntitledWorkspaceInfo, getStoredWorkspaceFolder, IEnterWorkspaceResult, isUntitledWorkspace, isWorkspaceIdentifier, ISingleFolderWorkspaceIdentifier } from 'vs/platform/workspaces/common/workspaces';
+import { toWorkspaceFolders, IWorkspaceIdentifier, hasWorkspaceFileExtension, UNTITLED_WORKSPACE_NAME, IResolvedWorkspace, IStoredWorkspaceFolder, isStoredWorkspaceFolder, IWorkspaceFolderCreationData, IUntitledWorkspaceInfo, getStoredWorkspaceFolder, IEnterWorkspaceResult, isUntitledWorkspace, isWorkspaceIdentifier, IStoredWorkspace } from 'vs/platform/workspaces/common/workspaces';
 import { IEnvironmentMainService } from 'vs/platform/environment/electron-main/environmentMainService';
 import { join, dirname } from 'vs/base/common/path';
 import { rimrafSync, readdirSync, writeFileSync, Promises } from 'vs/base/node/pfs';
-import { readFileSync, existsSync, mkdirSync, statSync, Stats } from 'fs';
-import { isLinux, isMacintosh, isWindows } from 'vs/base/common/platform';
+import { readFileSync, existsSync, mkdirSync } from 'fs';
+import { isWindows } from 'vs/base/common/platform';
 import { Event, Emitter } from 'vs/base/common/event';
 import { ILogService } from 'vs/platform/log/common/log';
-import { createHash } from 'crypto';
+import { getWorkspaceIdentifier } from 'vs/platform/workspaces/electron-main/workspaces';
 import { parse } from 'vs/base/common/json';
 import { URI } from 'vs/base/common/uri';
 import { Schemas } from 'vs/base/common/network';
@@ -54,12 +54,9 @@ export interface IWorkspacesManagementMainService {
 	isUntitledWorkspace(workspace: IWorkspaceIdentifier): boolean;
 
 	resolveLocalWorkspaceSync(path: URI): IResolvedWorkspace | undefined;
-	getWorkspaceIdentifier(workspacePath: URI): Promise<IWorkspaceIdentifier>;
-}
+	resolveLocalWorkspace(path: URI): Promise<IResolvedWorkspace | undefined>;
 
-export interface IStoredWorkspace {
-	folders: IStoredWorkspaceFolder[];
-	remoteAuthority?: string;
+	getWorkspaceIdentifier(workspacePath: URI): Promise<IWorkspaceIdentifier>;
 }
 
 export class WorkspacesManagementMainService extends Disposable implements IWorkspacesManagementMainService {
@@ -85,6 +82,16 @@ export class WorkspacesManagementMainService extends Disposable implements IWork
 	}
 
 	resolveLocalWorkspaceSync(uri: URI): IResolvedWorkspace | undefined {
+		return this.doResolveLocalWorkspace(uri, path => readFileSync(path, 'utf8'));
+	}
+
+	resolveLocalWorkspace(uri: URI): Promise<IResolvedWorkspace | undefined> {
+		return this.doResolveLocalWorkspace(uri, path => Promises.readFile(path, 'utf8'));
+	}
+
+	private doResolveLocalWorkspace(uri: URI, contentsFn: (path: string) => string): IResolvedWorkspace | undefined;
+	private doResolveLocalWorkspace(uri: URI, contentsFn: (path: string) => Promise<string>): Promise<IResolvedWorkspace | undefined>;
+	private doResolveLocalWorkspace(uri: URI, contentsFn: (path: string) => string | Promise<string>): IResolvedWorkspace | undefined | Promise<IResolvedWorkspace | undefined> {
 		if (!this.isWorkspacePath(uri)) {
 			return undefined; // does not look like a valid workspace config file
 		}
@@ -93,14 +100,16 @@ export class WorkspacesManagementMainService extends Disposable implements IWork
 			return undefined;
 		}
 
-		let contents: string;
 		try {
-			contents = readFileSync(uri.fsPath, 'utf8');
-		} catch (error) {
+			const contents = contentsFn(uri.fsPath);
+			if (contents instanceof Promise) {
+				return contents.then(value => this.doResolveWorkspace(uri, value), error => undefined /* invalid workspace */);
+			} else {
+				return this.doResolveWorkspace(uri, contents);
+			}
+		} catch {
 			return undefined; // invalid workspace
 		}
-
-		return this.doResolveWorkspace(uri, contents);
 	}
 
 	private isWorkspacePath(uri: URI): boolean {
@@ -115,7 +124,8 @@ export class WorkspacesManagementMainService extends Disposable implements IWork
 				id: workspaceIdentifier.id,
 				configPath: workspaceIdentifier.configPath,
 				folders: toWorkspaceFolders(workspace.folders, workspaceIdentifier.configPath, extUriBiasedIgnorePathCase),
-				remoteAuthority: workspace.remoteAuthority
+				remoteAuthority: workspace.remoteAuthority,
+				transient: workspace.transient
 			};
 		} catch (error) {
 			this.logService.warn(error.toString());
@@ -313,78 +323,4 @@ export class WorkspacesManagementMainService extends Disposable implements IWork
 
 		return { workspace, backupPath };
 	}
-}
-
-// !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-// NOTE: DO NOT CHANGE. IDENTIFIERS HAVE TO REMAIN STABLE
-// !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-
-export function getWorkspaceIdentifier(configPath: URI): IWorkspaceIdentifier {
-
-	function getWorkspaceId(): string {
-		let configPathStr = configPath.scheme === Schemas.file ? originalFSPath(configPath) : configPath.toString();
-		if (!isLinux) {
-			configPathStr = configPathStr.toLowerCase(); // sanitize for platform file system
-		}
-
-		return createHash('md5').update(configPathStr).digest('hex');
-	}
-
-	return {
-		id: getWorkspaceId(),
-		configPath
-	};
-}
-
-// !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-// NOTE: DO NOT CHANGE. IDENTIFIERS HAVE TO REMAIN STABLE
-// !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-
-export function getSingleFolderWorkspaceIdentifier(folderUri: URI): ISingleFolderWorkspaceIdentifier | undefined;
-export function getSingleFolderWorkspaceIdentifier(folderUri: URI, folderStat: Stats): ISingleFolderWorkspaceIdentifier;
-export function getSingleFolderWorkspaceIdentifier(folderUri: URI, folderStat?: Stats): ISingleFolderWorkspaceIdentifier | undefined {
-
-	function getFolderId(): string | undefined {
-
-		// Remote: produce a hash from the entire URI
-		if (folderUri.scheme !== Schemas.file) {
-			return createHash('md5').update(folderUri.toString()).digest('hex');
-		}
-
-		// Local: produce a hash from the path and include creation time as salt
-		if (!folderStat) {
-			try {
-				folderStat = statSync(folderUri.fsPath);
-			} catch (error) {
-				return undefined; // folder does not exist
-			}
-		}
-
-		let ctime: number | undefined;
-		if (isLinux) {
-			ctime = folderStat.ino; // Linux: birthtime is ctime, so we cannot use it! We use the ino instead!
-		} else if (isMacintosh) {
-			ctime = folderStat.birthtime.getTime(); // macOS: birthtime is fine to use as is
-		} else if (isWindows) {
-			if (typeof folderStat.birthtimeMs === 'number') {
-				ctime = Math.floor(folderStat.birthtimeMs); // Windows: fix precision issue in node.js 8.x to get 7.x results (see https://github.com/nodejs/node/issues/19897)
-			} else {
-				ctime = folderStat.birthtime.getTime();
-			}
-		}
-
-		// we use the ctime as extra salt to the ID so that we catch the case of a folder getting
-		// deleted and recreated. in that case we do not want to carry over previous state
-		return createHash('md5').update(folderUri.fsPath).update(ctime ? String(ctime) : '').digest('hex');
-	}
-
-	const folderId = getFolderId();
-	if (typeof folderId === 'string') {
-		return {
-			id: folderId,
-			uri: folderUri
-		};
-	}
-
-	return undefined; // invalid folder
 }
