+++ src/vs/workbench/browser/actions/workspaceCommands.ts
@@ -63,6 +63,7 @@ CommandsRegistry.registerCommand({
 		const workspaceEditingService = accessor.get(IWorkspaceEditingService);
 		const dialogsService = accessor.get(IFileDialogService);
 		const pathService = accessor.get(IPathService);
+
 		const folders = await dialogsService.showOpenDialog({
 			openLabel: mnemonicButtonLabel(localize({ key: 'add', comment: ['&& denotes a mnemonic'] }, "&&Add")),
 			title: localize('addFolderToWorkspaceTitle', "Add Folder to Workspace"),
@@ -117,7 +118,6 @@ CommandsRegistry.registerCommand(PICK_WORKSPACE_FOLDER_COMMAND_ID, async functio
 
 	const token: CancellationToken = (args ? args[1] : undefined) || CancellationToken.None;
 	const pick = await quickInputService.pick(folderPicks, options, token);
-
 	if (pick) {
 		return folders[folderPicks.indexOf(pick)];
 	}
@@ -138,6 +138,7 @@ CommandsRegistry.registerCommand({
 	id: 'vscode.openFolder',
 	handler: (accessor: ServicesAccessor, uri?: URI, arg?: boolean | IOpenFolderAPICommandOptions) => {
 		const commandService = accessor.get(ICommandService);
+
 		// Be compatible to previous args by converting to options
 		if (typeof arg === 'boolean') {
 			arg = { forceNewWindow: arg };
@@ -148,10 +149,12 @@ CommandsRegistry.registerCommand({
 			const options: IPickAndOpenOptions = {
 				forceNewWindow: arg?.forceNewWindow
 			};
+
 			if (arg?.forceLocalWindow) {
 				options.remoteAuthority = null;
 				options.availableFileSystems = ['file'];
 			}
+
 			return commandService.executeCommand('_files.pickFolderAndOpen', options);
 		}
 
@@ -163,6 +166,7 @@ CommandsRegistry.registerCommand({
 			noRecentEntry: arg?.noRecentEntry,
 			remoteAuthority: arg?.forceLocalWindow ? null : undefined
 		};
+
 		const uriToOpen: IWindowOpenable = (hasWorkspaceFileExtension(uri) || uri.scheme === Schemas.untitled) ? { workspaceUri: uri } : { folderUri: uri };
 		return commandService.executeCommand('_files.windowOpen', [uriToOpen], options);
 	},
@@ -198,11 +202,13 @@ interface INewWindowAPICommandOptions {
 CommandsRegistry.registerCommand({
 	id: 'vscode.newWindow',
 	handler: (accessor: ServicesAccessor, options?: INewWindowAPICommandOptions) => {
+		const commandService = accessor.get(ICommandService);
+
 		const commandOptions: IOpenEmptyWindowOptions = {
 			forceReuseWindow: options && options.reuseWindow,
 			remoteAuthority: options && options.remoteAuthority
 		};
-		const commandService = accessor.get(ICommandService);
+
 		return commandService.executeCommand('_files.newWindow', commandOptions);
 	},
 	description: {
@@ -225,16 +231,17 @@ CommandsRegistry.registerCommand('_workbench.removeFromRecentlyOpened', function
 	return workspacesService.removeRecentlyOpened([uri]);
 });
 
-
 CommandsRegistry.registerCommand({
 	id: 'vscode.removeFromRecentlyOpened',
 	handler: (accessor: ServicesAccessor, path: string | URI): Promise<any> => {
+		const workspacesService = accessor.get(IWorkspacesService);
+
 		if (typeof path === 'string') {
 			path = path.match(/^[^:/?#]+:\/\//) ? URI.parse(path) : URI.file(path);
 		} else {
 			path = URI.revive(path); // called from extension host
 		}
-		const workspacesService = accessor.get(IWorkspacesService);
+
 		return workspacesService.removeRecentlyOpened([path]);
 	},
 	description: {
@@ -254,10 +261,11 @@ interface RecentEntry {
 
 CommandsRegistry.registerCommand('_workbench.addToRecentlyOpened', async function (accessor: ServicesAccessor, recentEntry: RecentEntry) {
 	const workspacesService = accessor.get(IWorkspacesService);
-	let recent: IRecent | undefined = undefined;
 	const uri = recentEntry.uri;
 	const label = recentEntry.label;
 	const remoteAuthority = recentEntry.remoteAuthority;
+
+	let recent: IRecent | undefined = undefined;
 	if (recentEntry.type === 'workspace') {
 		const workspace = await workspacesService.getWorkspaceIdentifier(uri);
 		recent = { workspace, label, remoteAuthority };
@@ -266,10 +274,12 @@ CommandsRegistry.registerCommand('_workbench.addToRecentlyOpened', async functio
 	} else {
 		recent = { fileUri: uri, label, remoteAuthority };
 	}
+
 	return workspacesService.addRecentlyOpened([recent]);
 });
 
 CommandsRegistry.registerCommand('_workbench.getRecentlyOpened', async function (accessor: ServicesAccessor) {
 	const workspacesService = accessor.get(IWorkspacesService);
+
 	return workspacesService.getRecentlyOpened();
 });
